packages/claudomation/pyproject.toml
[project]
name = "claudomation"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "cyclopts>=3.10.1",
    "mmcp-client",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "textual>=2.1.2",
]

[project.scripts]
claudomation = "claudomation.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
mmcp-client = { workspace = true }
